+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/devices/huami/HuamiCoordinator.java
@@ -97,6 +97,9 @@ public abstract class HuamiCoordinator extends AbstractBLEDeviceCoordinator {
     }
 
     @Override
+    public boolean supportsFlashing() { return true; }
+
+    @Override
     public Class<? extends Activity> getAppsManagementActivity() {
         return null;
     }