+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/devices/pinetime/PineTimeJFCoordinator.java
@@ -62,6 +62,9 @@ public class PineTimeJFCoordinator extends AbstractBLEDeviceCoordinator {
     }
 
     @Override
+    public boolean supportsFlashing() { return true; }
+
+    @Override
     public boolean supportsActivityDataFetching() {
         return false;
     }